working fileInput examples for pictures and spreadsheets
addActionButtons returns a list $data and $button.colNames
addActionButtons separates the button name prefix from the row number with an underscore '_'
row number sent to callback.actionButton is now computed from the button name, rather than from the 'selected' row

+++ inst/examples/example_fileInput.R
@@ -1,38 +1,50 @@
-# minimal DTedit file input example
+# minimal DTedit file input example, using blobs
 \donttest{
   library(shiny)
   library(DTedit)
   library(blob)
-
+  
   server <- function(input, output) {
-
+    
+    picture <- reactiveVal(NULL)
+    spreadsheet <- reactiveVal(NULL)
+    
     my.actionButton.callback <- function(data, row, buttonID) {
-      ns <- parent.frame(1)$ns # the namespace of the calling environment
-      outfile <- tempfile(fileext = ".png")
-      zz <- file(outfile, "wb")
-      writeBin(object = unlist(data[row, "Picture"]), con = zz)
-      close(zz)
-      shiny::showModal(
-        shiny::modalDialog(
-          shiny::fluidPage(
-            shiny::tags$img(
-              src = base64enc::dataURI(file = outfile),
-              width = "100%"
-            )
-          ),
-          size = "l",
-          footer = list(
-            shiny::actionButton(ns("closeMyModal"), "OK")
-          )
-        )
-      )
-      browser()
-
-      # cleanup
-      file.remove(outfile)
+      if (substr(buttonID, 1, nchar("picture")) == "picture") {
+        if (length(unlist(data[row, "Picture"])) > 0) {
+          outfile <- tempfile(fileext = ".png")
+          # create temporary filename
+          zz <- file(outfile, "wb") # create temporary file
+          writeBin(object = unlist(data[row, "Picture"]), con = zz)
+          close(zz)
+          
+          picture(base64enc::dataURI(file = outfile))
+          
+          # cleanup
+          file.remove(outfile)
+        } else {
+          picture(NULL)
+        }
+      }
+      if (substr(buttonID, 1, nchar("spreadsheet")) == "spreadsheet") {
+        if (length(unlist(data[row, "Spreadsheet"])) > 0) {
+          outfile <- tempfile(fileext = ".csv")
+          # create temporary filename
+          zz <- file(outfile, "wb") # create temporary file
+          writeBin(object = unlist(data[row, "Spreadsheet"]), con = zz)
+          close(zz)
+          
+          spreadsheet(read.csv(outfile))
+          
+          # cleanup
+          file.remove(outfile)
+        } else {
+          spreadsheet(NULL)
+        }
+      }
       return(NULL)
     }
-
+    
     Grocery_List <- callModule(
       dtedit,
       'Grocery_List',
@@ -40,27 +52,54 @@
         Buy = c('Tea', 'Biscuits', 'Apples'),
         Quantity = c(7, 2, 5),
         Picture = c(as.blob(raw(0)), as.blob(raw(0)), as.blob(raw(0))),
+        Spreadsheet = c(as.blob(raw(0)), as.blob(raw(0)), as.blob(raw(0))),
         stringsAsFactors = FALSE
       ),
       view.cols = c("Buy", "Quantity"),
-      edit.cols = c("Buy", "Quantity", "Picture"),
-      input.choices = list(Picture = ".png"),
+      edit.cols = c("Buy", "Quantity", "Picture", "Spreadsheet"),
+      edit.label.cols = c(
+        "Item to buy", "Quantity",
+        "Picture (.png)", "Spreadsheet (.csv)"
+      ),
+      input.choices = list(Picture = ".png", Spreadsheet = ".csv"),
+      # unfortunately, RStudio's 'browser' doesn't actually respect
+      #  file-name/type restrictions. A 'real' browser does respect
+      #  the restrictions.
       action.button = list(
         MyActionButton = list(
-          columnLabel = "Show Picture",
-          buttonLabel = "Show",
-          buttonPrefix = "button_",
-          afterColumn = "Quantity")
+          columnLabel = "Picture",
+          buttonLabel = "Show Picture",
+          buttonPrefix = "picture",
+          afterColumn = "Quantity"),
+        MyOtherActionButton = list(
+          columnLabel = "Spreadsheet",
+          buttonLabel = "Show Spreadsheet",
+          buttonPrefix = "spreadsheet"
+        )
       ),
       callback.actionButton = my.actionButton.callback
     )
+    
+    output$listPicture <- shiny::renderUI({
+      shiny::tags$img(
+        src = picture(),
+        width = "100%"
+      )
+    })
+    
+    output$showSpreadsheet <- DT::renderDataTable({
+      spreadsheet()
+    })
   }
-
+  
   ui <- fluidPage(
-    h3('Grocery List'),
-    dteditUI('Grocery_List')
+    shinyalert::useShinyalert(),
+    h3("Grocery List"),
+    dteditUI("Grocery_List"),
+    uiOutput("listPicture"),
+    DT::dataTableOutput("showSpreadsheet")
   )
-
+  
   shinyApp(ui = ui, server = server)
 }
